requirements.txt
anyio==4.12.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
black==26.1.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-pkg==0.15
constantly==23.10.4
coverage==7.13.1
cryptography==46.0.5
daphne==4.2.1
Django==5.0.14
django-cors-headers==4.9.0
django-environ==0.12.1
django-filter==25.1
django-redis==6.0.0
django-unfold==0.77.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.14
exceptiongroup==1.3.1
gunicorn==25.1.0
h11==0.16.0
httptools==0.7.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
msgpack==1.1.2
mypy_extensions==1.1.0
ngrok==1.4.0
packaging==25.0
pathspec==1.0.4
pillow==12.0.0
platformdirs==4.9.2
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.10.1
pyngrok==7.5.0
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-django==4.11.1
python-click==0.1.4
python-decouple==3.8
python-dotenv==1.2.1
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
service-identity==24.2.0
sqlparse==0.5.5
stripe==14.1.0
tomli==2.4.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.3
unicorn==2.1.4
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
whitenoise==6.6.0
zope.interface==8.2
boto3